refactor: create IntlComputedProperty base class
