+++ addon/index.js
@@ -1,2 +1,3 @@
 export { default as Service } from './services/intl';
 export { default as translationMacro, raw } from './macro';
+export { default as IntlComputedProperty } from './intl-computed-property';
